JIT: use profile data even when schema doesn't match
After looking at some of the cases of pgo schema mismatch (#51908), the pgo
schemas still seem to have partial validity, so by default the jit will now
use the schema and data even for mismatches.

Add the ability to optionally assert on mismatches.

Also add a mode where the the jit compares the schema it provided to the
the runtime with the schema the runtime gives back to the jit, to try and
locate possible schema compression/decompression issues that might lead to the
sorts of mismatches we're seeing.

This turns out not to have found any issues. All the mismatches we see are
from schemas provided to the jit via static PGO. But it seems like a useful
capability to retain.