Enable partial compilation of methods
Enable partial compilation of Tier0 methods, using the simple strategy
that any stack-empty non-entry non-handler rarely run block is worth
deferring.

Since Tier0 curently does not access static profile data, this means
we mainly defer compilation of blocks with explicit throws.

This leverages patchpoints to trigger compilation of the missing parts
of methods in a manner similar to OSR. It is enabled for x64 and arm64.