JIT: clone loops with invariant indirs
Motivating example is the OSR loop in #78110 where in OSR we don't see
a deref before the loop, so can't hoist repeated virtual method lookups.

Turns out that these indirs are conditional so we end up not hoisting,
but at least this does the cloning part.

Still needs some polish, but should be functionally correct.